Refactor app-core jaxrs package
